font-access/font_access-chooser-selection.tentative.manual.https.html

<!doctype html>
<meta charset=utf-8>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/test-expectations.js"></script>

<script>
  promise_test(async t => {
    await new Promise(resolve => {
      window.addEventListener('DOMContentLoaded', resolve);
    });
    // Small delay to give chrome's test automation a chance to actually install
    // itself.
    await new Promise(resolve => step_timeout(resolve, 100));

    await window.test_driver.bless('show a font chooser.<br />Please select all the fonts that shows up.');
    // Arial is considered to be web-safe.
    let fonts = await navigator.fonts.showFontChooser({select: ['ArialMT']});
    assert_true(Array.isArray(fonts));
    assert_equals(fonts.length, 1);

    let postscriptName = fonts[0].postscriptName;
    assert_equals(typeof postscriptName, "string");
    assert_greater_than(postscriptName.length, 0);

    await window.test_driver.bless('show a font chooser.<br />Please select all the fonts that shows up.');
    fonts = await navigator.fonts.showFontChooser({select: []});
    assert_true(Array.isArray(fonts));
    assert_greater_than_equal(fonts.length, 1);

    postscriptName = fonts[0].postscriptName;
    assert_equals(typeof postscriptName, "string");
    assert_greater_than(postscriptName.length, 0);
  }, 'showFontChooser with selection works');
</script>
